requirements.txt
torch
torchvision
matplotlib
numpy<2
Pillow
tqdm
labelme
